+++ dashboard-app/src/features/dashboard/components/DataCard/types.ts
@@ -1,0 +1,23 @@
+
+/*
+ * Types
+ */
+
+export type TileDataPoint<T> = {
+  label: string
+  data: T
+};
+
+export type NumberTileProps = {
+  topText: string []
+  left: TileDataPoint<string[]>
+  right: TileDataPoint<number>
+  bottomText?: string []
+};
+
+export type TextTileProps = {
+  topText: string []
+  left: TileDataPoint<string[]>
+  right: TileDataPoint<string[]>
+  bottomText?: string []
+};
