+++ dashboard-app/src/lib/ui/index.css
@@ -1,4 +1,4 @@
-@import url('https://fonts.googleapis.com/css?family=Roboto:300,400');
+@import url('https://fonts.googleapis.com/css?family=Roboto:300,400,700');
 
 * {
   box-sizing: border-box;
